Refactor trackpad swipe overlay
